+++ src/chrome-polyfill.js
@@ -1,8 +1,12 @@
 if (typeof browser === "undefined") {
   if (chrome) {
     browser = chrome; // eslint-disable-line
-    const cbGet = browser.storage.local.get; // eslint-disable-line
+    const storageCbGet = browser.storage.local.get;
     browser.storage.local.get = keys =>
-      new Promise(resolve => cbGet(keys, resolve));
+      new Promise(resolve => storageCbGet(keys, resolve));
+
+    const tabCbGet = browser.tabs.get;
+    browser.tabs.get = tabId =>
+      new Promise(resolve => tabCbGet(tabId, resolve));
   }
 }